+++ backend/src/utils/eventScrape.ts
@@ -1,14 +1,13 @@
 import axios, { AxiosError } from 'axios';
 import { Event } from '../db/entities/Event';
-import { activityRepo } from '../repositories/Activity';
+import { activityRepo } from '../db/repositories/Activity';
 import { Activity } from '../db/entities/Activity';
 import * as cheerio from 'cheerio';
-import { eventRepo } from '../repositories/Event';
-import { arch } from 'os';
-import { ActivityModel } from '../models/Activity';
+import { eventRepo } from '../db/repositories/Event';
+import { ActivityModel } from '../db/models/Activity';
 import { Tournament } from '../db/entities/Tournament';
-import { tournamentRepo } from '../repositories/Tournament';
-import { EventModel } from '../models/Event';
+import { tournamentRepo } from '../db/repositories/Tournament';
+import { EventModel } from '../db/models/Event';
 
 function fetchPage(url: string): Promise<string|undefined> {
     const HTMLData =
